requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
Automat==24.8.1
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
cymem==2.0.11
defusedxml==0.7.1
Django==4.2.18
django-filter==24.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
exceptiongroup==1.2.2
fastapi==0.115.7
filelock==3.17.0
Flask==3.1.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gTTS==2.5.1
h11==0.14.0
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.5.0
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.1.1
Jinja2==3.1.5
jmespath==1.0.1
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.0
Mako==1.3.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.0.2
packaging==24.2
parsel==1.10.0
pillow==11.1.0
preshed==3.0.9
Protego==0.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.19.1
PyMySQL==1.1.0
pyOpenSSL==25.0.0
PyPDF2==3.0.1
queuelib==1.7.0
requests==2.31.0
requests-file==2.1.0
rich==13.9.4
Scrapy==2.12.0
service-identity==24.2.0
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
sqlparse==0.4.2
srsly==2.5.1
starlette==0.45.3
thinc==8.3.4
tldextract==5.1.3
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.0
uvicorn==0.34.0
w3lib==2.3.1
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
zope.interface==7.2
